+++ CHANGES.rst
@@ -1,7 +1,14 @@
 Changelog
 =========
 
-2.0.0 (unreleased)
+2.0.1 (unreleased)
+------------------
+
+- Fixed broken tests
+- Add LIS2-A2 Request Information capabilities
+
+
+2.0.0
 ------------------
 
 - Compatibility with SENAITE 2.x and senaite.astm